backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
google-cloud-bigquery==3.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
supervisor==4.2.5
drf-spectacular==0.26.5
whitenoise==6.5.0
psycopg2-binary==2.9.9
pandas==2.2.2